Move charity module to API
